requirements.txt
discord.py==1.3.3
google-api-python-client==1.8.2
psycopg2-binary==2.8.5